NP-46803-expand-channel-registry-journal, code-review
